requirements.txt
mariadb==1.0.9
mypy==0.920
py-algorand-sdk==1.9.0
python-dotenv==0.19.2